+++ bigwheel.js
@@ -185,6 +185,30 @@
         return this;
       }, // end bW.all
 
+      first : function (func, args) {
+        var args_array = [],
+            i;
+
+        // copy everything to args_array
+        // args, ^above^, should be an array-like object. if not, convert it.
+        if (!args.length) {
+          for (i = 1; i < arguments.length; i += 1) {
+            args_array.push(arguments[i]);
+          }
+        }
+        else {
+          for (i = 0; i < args.length; i += 1) {
+            args_array.push(args[i]);
+          }
+        }
+
+        args_array.unshift(this[0]);
+        func.apply(this, args_array);
+        args_array.shift();
+
+        return this; // ### TODO: return first element wrapped in bW wrapper ###
+      }, // end bW.first
+
       event_registry : { length : 0 },
 
       css : function (prop, value) {
@@ -268,7 +292,11 @@
         }
 
         return this.all(insert, arguments);
-      } // end bW.after
+      }, // end bW.after
+
+      val : function () { // returns a value, not the bW object
+        if (this[0].value) { return this[0].value; }
+      } // end bW.val
 
     } // end Bigwheel prototype